Created Cleaning Loop
Created a loop for cleaning, processing, and organizing the thermal image data.

This will then be plugged into the overarching Shiny app as a first step when thermal images are uploaded. After this step the image registration via an affine transformation would occur.

Ultimately I would like to potentially train the algorithm for use in the Shiny app, but we will see about that.