Huge Progress with Homography Matrix Creation
Succesfully created a couple of homography matrices, so far the image registration seems to going alright.

But there is a problem when resizing the registered image to the thermal image dimensions. When registered the image has too much black space and wouldn't actually correctly overlay in theory. Need to figure out how to correctly resize.

If this doesn't work then manual registration via tedious realignment from cropping and translation may be necessary.